Struct ic_crypto_internal_types::curves::bls12_381::G2

pub struct G2(pub [u8; 96]);

A point in the group "G2".

Content

bit (byte[0]>>7): compression flag: Should always be set to 1 as the 96 byte representation is compressed.

bit (byte[0]>>6): infinity: 1 if the point is at infinity, zero otherwise.

bit (byte[0]>>5): sort flag: In the compressed form, where only x is provided, this gives the sign of y.

first 48 bytes excluding flags (byte[0]&0x1f concat byte[1..=47]): The part "c1" of the x coordinate in big endian format.

next 48 bytes (byte[48..=95]): The part "c0" of the x coordinate in big endian format

Interoperability

This representation is consistent with IETF draft: https://datatracker.ietf.org/doc/draft-irtf-cfrg-bls-signature/?include_text=1 which specifies the "compressed serialization formats for E1 and E2 defined by [ZCash]"

This representation is different from the MIRACL library format, which stores the flag bits in a separate leading byte.

Ordering is by c1, or by c0 if the c1 are identical.

Miracl refers to c0 and c1 as a and b.

Implementations

impl G2

pub const SIZE: usize

pub const COMPRESSED_FLAG: u8

pub const INFINITY_FLAG: u8

pub const SIGN_FLAG: u8

pub const NON_FLAG_BITS: u8

pub const FLAG_BYTE_OFFSET: usize

pub const X1_BYTES_OFFSET: usize

pub const X0_BYTES_OFFSET: usize

pub fn as_bytes(&self) -> &[u8; 96]