Module ic_crypto_internal_types::sign::threshold_sig::ni_dkg::ni_dkg_groth20_bls12_381

Data types for the Groth20 non-interactive distributed key generation scheme.

Re-exports

pub use super::Epoch;

pub use crate::curves::bls12_381::Fr;

pub use crate::curves::bls12_381::G1;

pub use crate::curves::bls12_381::G2;

pub use crate::encrypt::forward_secure::groth20_bls12_381::Chunk;

pub use crate::encrypt::forward_secure::groth20_bls12_381::FsEncryptionCiphertext;

pub use crate::encrypt::forward_secure::groth20_bls12_381::FsEncryptionPlaintext;

pub use crate::encrypt::forward_secure::groth20_bls12_381::FsEncryptionPop;

pub use crate::encrypt::forward_secure::groth20_bls12_381::FsEncryptionPublicKey;

pub use crate::encrypt::forward_secure::groth20_bls12_381::NUM_CHUNKS;

pub use crate::sign::eddsa::ed25519::PublicKey;

pub use crate::sign::eddsa::ed25519::Signature;

pub use crate::sign::threshold_sig::public_coefficients::bls12_381::PublicCoefficientsBytes;

pub use crate::NodeIndex;

pub use FsEncryptionCiphertext as EncryptedShares;

Structs

Dealing
Threshold signature key material with proofs of correctness.

Transcript
All the public data needed for threshold key derivation.

ZKProofDec
A zero knowledge proof that the encrypted shares can be decrypted by the corresponding receivers.

ZKProofShare
A zero knowledge proof that the shares are indeed valid points on the curve.

Constants

NUM_ZK_REPETITIONS
The number of repetitions for zero-knowledge proofs.